requirements.txt
Pillow
instabot
requests